requirements.txt
imutils~=0.5.4
PyPDF2~=3.0.1
requests~=2.32.3
python-docx~=1.1.2
selenium~=4.27.1
arxiv~=2.1.3
google-api-python-client~=2.158.0
oauth2client~=4.1.3
beautifulsoup4~=4.12.3
python-dotenv~=0.19.1